Fix compiler warnings for cmake build
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@166674 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 381676df9cd6fbc097065740f9c0d9f004bf1e42

+++ Online/Online/Controller/FiniteStateMachine/Functors.h
@@ -132,7 +132,7 @@ namespace FiniteStateMachine {
     /// Slave mask
     int mask;
     /// Standard constructor. By default all slaves are acted on
-    SetSlaveState(const State* new_state, int match_mask=Slave::SLAVE_NONE) 
+  SetSlaveState(const State* new_state, int match_mask=(int)Slave::SLAVE_NONE) 
       : ActionCounter<State>(new_state), mask(match_mask) {}
     /// Operator invoked for each predicate to check if it is fulfilled
     void operator()(Slave* s);
